+++ Pandora/Paradigm/Primary/Functor/Equivalence.hs
@@ -4,14 +4,13 @@ import Pandora.Pattern.Category (($), (/))
 import Pandora.Pattern.Functor.Contravariant (Contravariant ((>$<)))
 import Pandora.Pattern.Functor.Divisible (Divisible ((>*<)))
 import Pandora.Pattern.Object.Ringoid (Ringoid ((*)))
-import Pandora.Paradigm.Primary.Object.Boolean (Boolean)
 import Pandora.Paradigm.Primary.Functor.Product (Product ((:*:)))
 
-data Equivalence a = Equivalence (a -> a -> Boolean)
+data Convergence r a = Convergence (a -> a -> r)
 
-instance Contravariant Equivalence where
-	f >$< Equivalence g = Equivalence $ \x y -> g / f x / f y
+instance Contravariant (Convergence r) where
+	f >$< Convergence g = Convergence $ \x y -> g / f x / f y
 
-instance Divisible Equivalence where
-	Equivalence g >*< Equivalence h = Equivalence $
+instance Ringoid r => Divisible (Convergence r) where
+	Convergence g >*< Convergence h = Convergence $
 		\(x :*: x') (y :*: y') -> g x y * h x' y'