+++ Pandora/Paradigm/Primary/Transformer/Flip.hs
@@ -2,7 +2,7 @@ module Pandora.Paradigm.Primary.Transformer.Flip where
 
 import Pandora.Pattern.Category (($))
 import Pandora.Pattern.Functor.Bivariant (Bivariant ((<->)))
-import Pandora.Paradigm.Controlflow.Effect.Interpreted (Interpreted (Primary, run, unite))
+import Pandora.Paradigm.Controlflow.Effect.Interpreted (Interpreted (Primary, run, unite, (||=)))
 
 newtype Flip (v :: * -> * -> *) a e = Flip (v e a)
 
@@ -12,4 +12,4 @@ instance Interpreted (Flip v a) where
 	unite = Flip
 
 instance Bivariant v => Bivariant (Flip v) where
-	f <-> g = \x -> unite $ g <-> f $ run x
+	f <-> g = \x -> (g <-> f) ||= x