+++ Pandora/Paradigm/Structure/Interface/Dictionary.hs
@@ -1,6 +1,1 @@
 module Pandora.Paradigm.Structure.Interface.Dictionary where
-
-import Pandora.Paradigm.Primary.Functor.Maybe (Maybe)
-
-class Dictionary a k t where
-	(?=) :: k -> t a -> Maybe a