+++ Pandora/Paradigm/Structure/Ability/Morphable.hs
@@ -56,9 +56,3 @@ filter p xs = run / morph @(Delete f) xs / p
 
 find :: forall f t u a . (Morphable (Find f) t, Morphing (Find f) t ~ (Predicate <:.:> u := (->))) => Predicate a -> t a -> u a
 find p xs = run / morph @(Find f) xs / p
-
-lookup :: forall f k t u a . (Morphable (Lookup f) t, Morphing (Lookup f) t ~ ((->) k <:.> u)) => k -> t a -> u a
-lookup key xs = run / morph @(Lookup f) xs / key
-
-discover :: forall f k v t u a . (Morphable (Lookup f) t, Morphing (Lookup f) t ~ ((->) (v k) <:.> u)) => v k -> t a -> u a
-discover keys xs = run / morph @(Lookup f) xs / keys