+++ CHANGELOG.md
@@ -384,3 +384,4 @@
 * Generalize `Equivalence` and rename it to `Convergence`
 * Define `Equivalence` and `Comparison` as type synonyms for `Convergence`
 * Define version of `<*>` with flipped arguments - `<%>`
+* Generalize `||=` and `=||` methods of `Interpreted` on another `Interpeted` functor