+++ CHANGELOG.md
@@ -393,3 +393,4 @@
 * Remove `Dictonary` typeclass
 * Define `Lookup` verb for finding elements in datastructures by key
 * Define `Cycle` typeclass with `Chain` superclass
+* Move `lookup` and `discover` exressions to `Dictionary` module